requirements.txt
anyio==3.6.2
asn1crypto==1.5.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
pg8000==1.29.4
pydantic==1.10.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.28.2
scramp==1.4.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
starlette==0.25.0
typing-extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
